% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_simple.R
\name{theme_simple}
\alias{theme_simple}
\title{A simple ggplot theme}
\usage{
theme_simple(base_size = 11, base_family = "", grid = FALSE)
}
\arguments{
\item{base_size}{Base text size}

\item{base_family}{Base font family}

\item{grid}{Logical, should major grid lines be drawn}
}
\description{
A simple theme for ggplot with thin boxes around plot regions
    and clean facet labels
}
\examples{
library(ggplot2)
g <- ggplot(mpg) +
    aes(x = displ, y = hwy) +
    geom_point(color = "grey40") +
    geom_smooth(method = "loess", color = "red3", span = 1, se = FALSE) +
    facet_wrap( ~ drv, nrow = 1)

g + theme_simple()
g + theme_simple(grid = TRUE)
}
